# facturation-planning-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour gérer la facturation et les plannings",
        "title": "Facturation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/entreprises": {
            "get": {
                "description": "Retourne la liste complète de toutes les entreprises",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entreprises"
                ],
                "summary": "Récupérer toutes les entreprises",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Entreprise"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Crée une nouvelle entreprise avec les informations fournies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entreprises"
                ],
                "summary": "Créer une nouvelle entreprise",
                "parameters": [
                    {
                        "description": "Informations de l'entreprise",
                        "name": "entreprise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Entreprise"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Entreprise"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/factures": {
            "get": {
                "description": "Retourne la liste complète des factures en base de données",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factures"
                ],
                "summary": "Lister toutes les factures",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Facture"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Ajoute une nouvelle facture en base de données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factures"
                ],
                "summary": "Créer une facture",
                "parameters": [
                    {
                        "description": "Détails de la facture",
                        "name": "facture",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Facture"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Facture"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/factures/{id}/pdf": {
            "get": {
                "description": "Génère un PDF pour une facture donnée",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Factures"
                ],
                "summary": "Récupérer une facture en PDF",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la facture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Facture non trouvée",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération du PDF",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/plannings": {
            "get": {
                "description": "Retourne la liste complète des plannings enregistrés",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planning"
                ],
                "summary": "Récupérer tous les plannings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Planning"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Ajoute un nouveau planning en base de données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planning"
                ],
                "summary": "Créer un planning",
                "parameters": [
                    {
                        "description": "Détails du planning",
                        "name": "planning",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Planning"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Planning"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/plannings/{id}": {
            "put": {
                "description": "Met à jour un planning existant par son ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Modifier un planning",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID du planning",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles informations du planning",
                        "name": "planning",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Planning"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Planning"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Planning non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Supprime un planning par son ID",
                "summary": "Supprimer un planning",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID du planning",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Planning non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Client": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string",
                    "example": "123 rue de la Paix, 75001 Paris"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "contact@acme.com"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "ACME Corporation"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                }
            }
        },
...
        "models.Entreprise": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string",
                    "example": "123 rue du Code"
                },
                "bic": {
                    "type": "string",
                    "example": "BIC12345"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "contact@maboite.com"
                },
                "iban": {
                    "type": "string",
                    "example": "FR7612345987650123456789014"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "MaBoite"
                },
                "responsable": {
                    "type": "string",
                    "example": "Jean Dupont"
                },
                "siret": {
                    "type": "string",
                    "example": "12345678901234"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "tva": {
                    "type": "string",
                    "example": "FR12345678901"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                }
            }
        },
        "models.Facture": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                },
                "date_echeance": {
                    "type": "string",
                    "example": "2025-04-17"
                },
                "date_emission": {
                    "type": "string",
                    "example": "2025-03-17"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Facture pour prestation de services"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "montant_ht": {
                    "type": "number",
                    "example": 1000.5
                },
                "montant_ttc": {
                    "type": "number",
                    "example": 1200.6
                },
                "numero": {
                    "type": "string",
                    "example": "FAC-2025-001"
                },
                "planning_id": {
                    "description": "Lien avec le planning si applicable",
                    "type": "integer",
                    "example": 3
                },
                "statut": {
                    "description": "\"payée\", \"en attente\", \"rejetée\"",
                    "type": "string",
                    "example": "en attente"
                },
                "taux_tva": {
                    "type": "number",
                    "example": 20
                },
                "type": {
                    "description": "\"standard\", \"etat_avancement\", \"auto_liquidation\"",
                    "type": "string",
                    "example": "standard"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                }
            }
        },
...
        "models.Planning": {
            "type": "object",
            "properties": {
                "client": {
                    "description": "\u003c-- ajoute cette ligne",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Client"
                        }
                    ]
                },
                "client_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                },
                "date": {
                    "type": "string",
                    "example": "2025-06-15"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 1
                },
                "facturation": {
                    "type": "string",
                    "example": "À facturer"
                },
                "facture": {
                    "$ref": "#/definitions/models.Facture"
                },
                "forfait_ht": {
                    "type": "number",
                    "example": 0
                },
                "heure_debut": {
                    "type": "string",
                    "example": "09:00"
                },
                "heure_fin": {
                    "type": "string",
                    "example": "17:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nb_repetitions": {
                    "description": "combien de fois on le répète",
                    "type": "integer"
                },
                "objet": {
                    "type": "string",
                    "example": "Maintenance système"
                },
                "periodicite": {
                    "type": "integer",
                    "example": 0
                },
                "prestation": {
                    "type": "string",
                    "example": "Support technique"
                },
                "salarie": {
                    "description": "✅ Corrigé",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Salarie"
                        }
                    ]
                },
                "salarie_id": {
                    "type": "integer",
                    "example": 1
                },
                "taux_horaire": {
                    "type": "number",
                    "example": 45.5
                },
                "type_evenement": {
                    "type": "string",
                    "example": "Intervention"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                }
            }
        },
...
        "models.Salarie": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "jean.dupont@example.com"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "Jean Dupont"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                }
            }
        }
    }
}